Terrain ID
Removed feature to generate a terrain object: the terrain data does not match the default terrain object created by the Unity Editor and it is easier to use the default, premade terrain.

begun creating methods to generate objects on terrains: created various ways to identify terrain objects and how to identify the current Y coordinate of the terrain and a given X and Z coordinate.